Refactor auth for modularity
