+++ .gitignore
@@ -7,11 +7,7 @@
 __debug_bin
 ui-server
 
-# Test binary, built with `go test -c`
-*.test
-
-# Output of the go coverage tool, specifically when used with LiteIDE
-*.out
+.vscode/*.log
 
 # Output of code generation tools
 third_party/OpenAPI/temporal
