+++ server/routes/auth.go
@@ -36,22 +36,11 @@ import (
 	"github.com/gorilla/sessions"
 	"github.com/labstack/echo-contrib/session"
 	"github.com/labstack/echo/v4"
+	"github.com/temporalio/ui-server/server/auth"
 	"github.com/temporalio/ui-server/server/config"
 	"golang.org/x/net/context"
 	"golang.org/x/oauth2"
 )
-
-type User struct {
-	OAuth2Token *oauth2.Token
-	IDToken     *Claims
-}
-
-type Claims struct {
-	Email         string `json:"email"`
-	EmailVerified bool   `json:"email_verified"`
-	Name          string `json:"name"`
-	Picture       string `json:"picture"`
-}
 
 // SetAuthRoutes sets routes used by auth
 func SetAuthRoutes(e *echo.Echo, cfgProvider *config.ConfigProviderWithRefresh) {
@@ -132,32 +121,23 @@ func authenticate(config *oauth2.Config, options map[string]interface{}) func(ec
 
 func authenticateCb(ctx context.Context, config *oauth2.Config, provider *oidc.Provider) func(echo.Context) error {
 	return func(c echo.Context) error {
-		user, err := exchangeCode(ctx, c.Request(), config, provider)
+		user, err := auth.ExchangeCode(ctx, c.Request(), config, provider)
 		if err != nil {
-			return err
+			return echo.NewHTTPError(http.StatusInternalServerError, "unable to exchange code", err)
 		}
 
-		sess, _ := session.Get("auth", c)
-		sess.Options = &sessions.Options{
-			Path:     "/",
-			MaxAge:   7 * 24 * int(time.Hour.Seconds()),
-			HttpOnly: true,
-			SameSite: http.SameSiteStrictMode,
-			Secure:   true,
+		err = auth.SetUser(c, user)
+		if err != nil {
+			return echo.NewHTTPError(http.StatusInternalServerError, "unable to set user", err)
 		}
-		sess.Values["access-token"] = &user.OAuth2Token.AccessToken
-		sess.Values["email"] = &user.IDToken.Email
-		sess.Values["picture"] = &user.IDToken.Picture
-		sess.Values["name"] = &user.IDToken.Name
-		sess.Save(c.Request(), c.Response())
 
 		nonceS, err := c.Request().Cookie("nonce")
 		if err != nil {
-			return echo.NewHTTPError(http.StatusBadRequest, "Nonce is not provided")
+			return echo.NewHTTPError(http.StatusBadRequest, "nonce is not provided")
 		}
 		nonce, err := nonceFromString(nonceS.Value)
 		if err != nil {
-			return echo.NewHTTPError(http.StatusBadRequest, "Nonce is invalid")
+			return echo.NewHTTPError(http.StatusBadRequest, "nonce is invalid")
 		}
 
 		returnUrl := nonce.ReturnURL
@@ -170,7 +150,7 @@ func authenticateCb(ctx context.Context, config *oauth2.Config, provider *oidc.P
 }
 
 func logout(c echo.Context) error {
-	sess, _ := session.Get("auth", c)
+	sess, _ := session.Get(auth.AuthCookie, c)
 	sess.Options = &sessions.Options{
 		Path:     "/",
 		MaxAge:   -1,
@@ -186,55 +166,6 @@ func logout(c echo.Context) error {
 	}
 
 	return c.Redirect(http.StatusSeeOther, returnUrl)
-}
-
-func exchangeCode(ctx context.Context, r *http.Request, config *oauth2.Config, provider *oidc.Provider) (*User, error) {
-	state, err := r.Cookie("state")
-	if err != nil {
-		return nil, echo.NewHTTPError(http.StatusBadRequest, "State cookie is not set in request")
-	}
-	if r.URL.Query().Get("state") != state.Value {
-		return nil, echo.NewHTTPError(http.StatusBadRequest, "State cookie did not match")
-	}
-
-	oauth2Token, err := config.Exchange(ctx, r.URL.Query().Get("code"))
-	if err != nil {
-		return nil, echo.NewHTTPError(http.StatusInternalServerError, "Unable to exchange token: "+err.Error())
-	}
-
-	rawIDToken, ok := oauth2Token.Extra("id_token").(string)
-	if !ok {
-		return nil, echo.NewHTTPError(http.StatusInternalServerError, "No id_token field in oauth2 token.")
-	}
-	oidcConfig := &oidc.Config{
-		ClientID: config.ClientID,
-	}
-	verifier := provider.Verifier(oidcConfig)
-	idToken, err := verifier.Verify(ctx, rawIDToken)
-	if err != nil {
-		return nil, echo.NewHTTPError(http.StatusInternalServerError, "Unable to verify ID Token: "+err.Error())
-	}
-
-	nonce, err := r.Cookie("nonce")
-	if err != nil {
-		return nil, echo.NewHTTPError(http.StatusBadRequest, "Nonce is not provided")
-	}
-	if idToken.Nonce != nonce.Value {
-		return nil, echo.NewHTTPError(http.StatusBadRequest, "Nonce did not match")
-	}
-
-	var claims Claims
-
-	if err := idToken.Claims(&claims); err != nil {
-		return nil, echo.NewHTTPError(http.StatusInternalServerError, err.Error())
-	}
-
-	user := User{
-		OAuth2Token: oauth2Token,
-		IDToken:     &claims,
-	}
-
-	return &user, nil
 }
 
 func setCallbackCookie(c echo.Context, name, value string) {
